[TEC-3339] Removed FlowcommerceSpree::Experience - Zone will be used instead
Removed FlowcommerceSpree::Setting ActiveRecord model - instead will use FlowcommerceSpree::Settings inheriting from Spree::Preferences::Configuration.

Add preference `:product_catalog_upload, :hash, default: {}` to Spree::Preferences::Configuration.

Add `:import_flowcommerce` method to Spree::Zone.

ApiRefresh namespaced under FlowcommerceSpree and refactored to work with FlowcommerceSpree::Config.product_catalog_upload

FlowcommerceSpree::Config is defined in Engine's `after_initialize` callback.
