+++ app/models/flowcommerce_spree/settings.rb
@@ -1,0 +1,5 @@
+module FlowcommerceSpree
+  class Settings < Spree::Preferences::Configuration
+    preference :product_catalog_upload, :hash, default: {}
+  end
+end
